Update PortablePanel.cs
Added `_ToggleTabOnHold`, which allows desktop users to toggle the tab behavior at runtime.

Added Constrained options for View and Position. This lets VR users `_ToggleViewConstrained` to lock the panel to their view or `_TogglePositionConstrained` to pin it to their local space. In both cases, the menu sticks to the player if they move around, but can still be grabbed/moved/resized.

Added `_ToggleLocked` so the panel can be locked in place. When enabled, it can't be grabbed, moved or resized, but this will be reset if it's closed or respawned.

There's also some commented out code from an attempt at making it push/pull-able with the joystick when held, but I couldn't get it working.